+++ extension/CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [1.0.0-beta-1] - 2026-03-10
+
+### Added
+
+- Setting: ATC.addMissingProperties
+
+### Changed
+
+- Renamed "ATC.TranslationMethod" to "ATC.translationMethod" for consistency
+  -Renamed "ATC: Find Translations in AL File" to "ATC: Write Translations to Comments in current AL File" for better understanding
+- Renamed "ATC: Find Translations in XLF File" to "ATC: Write Translations in current Xliff File to AL File Comments" for better understanding
+
+### Fixed
+
+- Comments have to be structured like this: '%1="Description", DEU="Translation", Other Comment'
+
+## [1.0.0-beta] - 2026-03-10
+
+### Added
+
+- Command: Find Translations in Xliff File (See [README.md](README.md#atc-find-translations-in-xliff-file) for more info)
+
 ## [0.0.1-beta] - 2026-03-09
 
 ### Added
@@ -11,10 +33,3 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 - Command: Find Translations in AL File (See [README.md](README.md#atc-find-translations-in-file) for more info)
 - Setting: ATC.TranslationMethod
 - Setting: ATC.languageMapping
-
-
-## [1.0.0-beta] - 2026-03-10
-
-### Added
-
-- Command: Find Translations in Xliff File (See [README.md](README.md#atc-find-translations-in-xliff-file) for more info)